fix: morph keywords around when updating meta
This is ugly, because the whole loop is unclear.

It initialises keywords as objects (as per the database), but it needs a string in the input (select works like that...). Then it updates correctly when it's a string, correcting it to an array for the mutations by itself.

We had to init the keywords value as a string in react, so it works with the select component. And then, morph all types throughout to reflect the fact that the the type changes here and there.