fix: update new order algorithm to exclude current question and lost artefacts
If you start to create a question, then refresh the page (instead of cancelling properly), a question is left, even if empty. It was used in the computation for order, even if the question was actually not in the page, thus creating wrong computations of indexes (on a very specific case).